refactor: rename ChatShortCut to ChatShortcut and update import paths
