+++ src/stories/chat/parts/shortcuts/index.stories.tsx
